refactor(FavoriteLocation): Extract FavoriteLocation into file.
