+++ lib/components/form/favorite-location-field.js
@@ -29,7 +29,7 @@ const StyledLocationField = styled(LocationField)`
     display: table-cell;
     font-size: 100%;
     line-height: 20px;
-    padding: 0;
+    padding: 6px 12px;
     width: 100%;
   }
 `
